PR comment: Make AccountLinkedViewModel separate
